+++ src/viewdom/__init__.py
@@ -1,6 +1,6 @@
 from viewdom.h import html, render, use_context, Context, VDOM
 
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 
 __all__ = [
     'html',